Keep tx data as Uint8Array and defer data_root computation
This changes Transaction so it no longer converts data to b64
during createTransaction, instead always keeping as Uint8Array.
It also defers the compution of data_root until you call
getSignatureData on the transaction object.

To maintain compat, we encode the Uint8Array back into base64url
when you call `tx.get('data')` and in the toJSON() method, so the the
regular `transactions.post()` method still works and the tx.get('field')
behaviour remains the same.