Add a passing and a failing test
The data_root computation does not match the way it is currently performed by the network nodes (v43), when
the last chunk is smaller than 32 KiB after splitting the preceding data into 256 KiB chunks.

The chunking method was updated in 01386257. It is a better way to chunk data and it should be continued to be used
in chunked uploads. However, when data is directly uploaded with the header, the exact same chunking scheme
the nodes use needs to be used.